+++ src/main/java/com/devsu/account/entity/dto/AccountRecord.java
@@ -1,4 +1,8 @@
 package com.devsu.account.entity.dto;
 
 public record AccountRecord(
-    String number, String clientIdentification, String type, Long initialBalance, Boolean status) {}
+    String number,
+    String clientIdentification,
+    String type,
+    Double initialBalance,
+    Boolean status) {}
